+++ library/src/schemas/instance/index.ts
@@ -1,2 +1,3 @@
 export * from './instance.ts';
 export * from './instanceAsync.ts';
+export * from './types.ts';