+++ website/src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './disableTransitions';
+export * from './addArticleMetadata';
 export * from './trackEvent';