+++ library/src/regex.ts
@@ -79,7 +79,7 @@ export const ISO_TIME_SECOND_REGEX = /^(?:0\d|1\d|2[0-3])(?::[0-5]\d){2}$/u;
  * [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp regex.
  */
 export const ISO_TIMESTAMP_REGEX =
-  /^\d{4}-(?:0[1-9]|1[0-2])-(?:[12]\d|0[1-9]|3[01])T(?:0\d|1\d|2[0-3])(?::[0-5]\d){2}\.\d{3,9}Z$/u;
+  /^\d{4}-(?:0[1-9]|1[0-2])-(?:[12]\d|0[1-9]|3[01])T(?:0\d|1\d|2[0-3])(?::[0-5]\d){2}(?:\.\d{1,9})?Z$/u;
 
 /**
  * [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) week regex.