+++ library/src/validations/isoTimestamp/isoTimestamp.test.ts
@@ -12,6 +12,8 @@ describe('isoTimestamp', () => {
     expect(validate._parse(value3).output).toBe(value3);
     const value4 = '2024-01-04T17:40:21.157953900Z';
     expect(validate._parse(value4).output).toBe(value4);
+    const value5 = '2024-01-16T16:00:34Z';
+    expect(validate._parse(value5).output).toBe(value5);
 
     expect(validate._parse('').issues).toBeTruthy();
     expect(validate._parse('2023-07-11T17:26:27.243').issues).toBeTruthy();
@@ -25,6 +27,10 @@ describe('isoTimestamp', () => {
     expect(validate._parse('0000-01-01T01:60:00.000Z').issues).toBeTruthy();
     expect(validate._parse('0000-01-01T01:00:60.000Z').issues).toBeTruthy();
     // FIXME: expect(validate._parse('2023-06-31T00:00:00.000Z').issues).toBeTruthy();
+    expect(validate._parse('0000-01-01T00:00:00.Z').issues).toBeTruthy();
+    expect(
+      validate._parse('0000-01-01T00:00:00.0000000000Z').issues
+    ).toBeTruthy();
   });
 
   test('should return custom error message', () => {