+++ library/src/schemas/union/union.ts
@@ -2,12 +2,14 @@ import type {
   BaseSchema,
   ErrorMessage,
   Input,
-  Issues,
   MaybeReadonly,
   Output,
   Pipe,
+  TypedSchemaResult,
+  UntypedSchemaResult,
 } from '../../types/index.ts';
 import { defaultArgs, pipeResult, schemaIssue } from '../../utils/index.ts';
+import { subissues } from './utils/index.ts';
 
 /**
  * Union options type.
@@ -36,7 +38,7 @@ export type UnionSchema<
   /**
    * The validation and transformation pipeline.
    */
-  pipe: Pipe<Input<TOptions[number]>> | undefined;
+  pipe: Pipe<Output<TOptions[number]>> | undefined;
 };
 
 /**
@@ -49,7 +51,7 @@ export type UnionSchema<
  */
 export function union<TOptions extends UnionOptions>(
   options: TOptions,
-  pipe?: Pipe<Input<TOptions[number]>>
+  pipe?: Pipe<Output<TOptions[number]>>
 ): UnionSchema<TOptions>;
 
 /**
@@ -64,13 +66,13 @@ export function union<TOptions extends UnionOptions>(
 export function union<TOptions extends UnionOptions>(
   options: TOptions,
   message?: ErrorMessage,
-  pipe?: Pipe<Input<TOptions[number]>>
+  pipe?: Pipe<Output<TOptions[number]>>
 ): UnionSchema<TOptions>;
 
 export function union<TOptions extends UnionOptions>(
   options: TOptions,
-  arg2?: Pipe<Input<TOptions[number]>> | ErrorMessage,
-  arg3?: Pipe<Input<TOptions[number]>>
+  arg2?: Pipe<Output<TOptions[number]>> | ErrorMessage,
+  arg3?: Pipe<Output<TOptions[number]>>
 ): UnionSchema<TOptions> {
   // Get message and pipe argument
   const [message = 'Invalid type', pipe] = defaultArgs(arg2, arg3);
@@ -83,39 +85,78 @@ export function union<TOptions extends UnionOptions>(
     message,
     pipe,
     _parse(input, info) {
-      // Create issues and output
-      let issues: Issues | undefined;
-      let output: [Output<TOptions[number]>] | undefined;
+      // Create variables to collect results
+      let validResult: TypedSchemaResult<Output<TOptions[number]>> | undefined;
+      let untypedResults: UntypedSchemaResult[] | undefined;
+      let typedResults:
+        | TypedSchemaResult<Output<TOptions[number]>>[]
+        | undefined;
 
       // Parse schema of each option
       for (const schema of this.options) {
         const result = schema._parse(input, info);
 
-        // If there are issues, capture them
-        if (result.issues) {
-          if (issues) {
-            for (const issue of result.issues) {
-              issues.push(issue);
-            }
+        // If typed, add valid or typed result
+        if (result.typed) {
+          // If there are no issues, add valid result and break loop
+          if (!result.issues) {
+            validResult = result;
+            break;
+
+            // Otherwise, add typed result
           } else {
-            issues = result.issues;
+            typedResults
+              ? typedResults.push(result)
+              : (typedResults = [result]);
           }
 
-          // Otherwise, set output and break loop
+          // Otherwise, add untyped result
         } else {
-          // Note: Output is nested in array, so that also a falsy value
-          // further down can be recognized as valid value
-          output = [result.output];
-          break;
+          untypedResults
+            ? untypedResults.push(result)
+            : (untypedResults = [result]);
         }
       }
 
-      // If there is an output, execute pipe
-      if (output) {
-        return pipeResult(output[0], this.pipe, info, 'union');
+      // If there is a valid result, execute pipe
+      if (validResult) {
+        return pipeResult(validResult.output, this.pipe, info, 'union');
+      }
+
+      // If there are typed results, execute pipe
+      if (typedResults?.length) {
+        const firstResult = typedResults[0];
+        return pipeResult(
+          firstResult.output,
+          this.pipe,
+          info,
+          'union',
+          // Hint: If there is more than one typed result, we use a general
+          // union issue with subissues because the issues could contradict
+          // each other.
+          typedResults.length === 1
+            ? firstResult.issues
+            : schemaIssue(
+                info,
+                'union',
+                'union',
+                this.message,
+                input,
+                undefined,
+                subissues(typedResults)
+              ).issues
+        );
+      }
+
+      // If there is only one untyped result, return it
+      if (untypedResults?.length === 1) {
+        return untypedResults[0];
       }
 
       // Otherwise, return schema issue
+      // Hint: If there are zero or more than one untyped results, we use a
+      // general union issue with subissues because the issues could contradict
+      // each other.
       return schemaIssue(
         info,
         'type',
@@ -123,7 +164,7 @@ export function union<TOptions extends UnionOptions>(
         this.message,
         input,
         undefined,
-        issues
+        subissues(untypedResults)
       );
     },
   };