+++ library/CHANGELOG.md
@@ -6,10 +6,10 @@ All notable changes to the library will be documented in this file.
 
 - Improve performance of `enum_` and `enumAsync` schema by caching values
 - Change ISO timestamp regex to support timestamps with lower and higher millisecond accuracy (pull request #353)
-- Change issue handling of `union` and `unionAsync` to improve developer experience
+- Change issue handling of `union`, `unionAsync`, `variant` and `variantAsync` to improve developer experience
 - Fix bug in `getDefaults`, `getDefaultsAsync`, `getFallbacks` and `getFallbacksAsync` for falsy but not `undefined` values (issue #356)
-- Fix type of `pipe` argument at `union` and `unionAsync`
-- Fix bug that broke pipeline execution in `union` and `unionAsync` (issue #364)
+- Fix type of `pipe` argument at `union`, `unionAsync`, `variant` and `variantAsync`
+- Fix bug that broke pipeline execution in `union`, `unionAsync`, `variant` and `variantAsync` (issue #364)
 
 ## v0.25.0 (December 26, 2023)