+++ library/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the library will be documented in this file.
 
-## vX.X.X (Month DD, YYYY)
+## v0.26.0 (January 16, 2023)
 
 - Improve performance of `enum_` and `enumAsync` schema by caching values
 - Change ISO timestamp regex to support timestamps with lower and higher millisecond accuracy (pull request #353)
@@ -10,6 +10,7 @@ All notable changes to the library will be documented in this file.
 - Fix bug in `getDefaults`, `getDefaultsAsync`, `getFallbacks` and `getFallbacksAsync` for falsy but not `undefined` values (issue #356)
 - Fix type of `pipe` argument at `union`, `unionAsync`, `variant` and `variantAsync`
 - Fix bug that broke pipeline execution in `union`, `unionAsync`, `variant` and `variantAsync` (issue #364)
+- Fix typo in type name of `startsWith` validation action (pull request #375)
 
 ## v0.25.0 (December 26, 2023)